Move logout to logout.py
