+++ APITaxi_front/views/admin.py
@@ -45,15 +45,3 @@ blueprint.add_url_rule(
     view_func=AdminLogAs.as_view('logas'),
     methods=['GET', 'POST']
 )
-
-
-class Logout(LogoutAsView):
-    user_model = User
-    redirect_on_success = 'home.home'
-
-
-blueprint.add_url_rule(
-    '/logas/logout',
-    view_func=Logout.as_view('logas_logout'),
-    methods=['POST']
-)
